[Refactor] UserInfo CloudKit 의존성 제거ã 및 Extension
